+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0]
+
+- added `lazyLoading` attribute to `<khao-image>`
+- removed `lazyloadSrc` and `lazyloadThreshold` from `<khao-image>` (/!\)
+
 ## [1.6.7]
 
 - upgraded to `khao-maletv0.0.13`